+++ src/components/commons/column-type/room.js
